backend-python/auth_service/requirements.txt
aiosqlite==0.21.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
coverage==7.8.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.3.4
pytest-asyncio==1.0.0
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.38.6
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.31.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
